Don't worry about CLOSURE and CLOS in Rebmu
Though the implementation debates of the "thing called
closure" vs. "the thing called function" go on, the question of
whether or not there will only be one type of function in the
distribution aren't relevant to Rebmu.  The semantics of
closure are the correct ones, and the only reason that
function exists is for a reduced behavior for performance.

Rebmu will call that one thing "FN" because it's the more
familiar word in programming culture...and also because it
frees up CL to use with the very useful COLLECT.