GC if/either/unless/while equality test compression
In Rebmu's early formulation, the "mu library" played a stronger role
since Rebol's IF/UNLESS/EITHER would not be able to use values
directly as conditions.  This opened the door to generally making
variants like EE to mean `either equals?` or EU to mean
`either not equals?` ("either unequal").

Feedback into the language and development pushed it to where
IF/UNLESS/EITHER changed, and the variations are now only
saving one character.  `eE?` as `either equals?` for instance.  The
migration toward teaching tool as well as seeking to win the
"end game" of longer code golf challenges makes such
optimizations less relevant to be including.