Kill only-first/second-true, index?-find-mu
Gets rid of some basically unused mu-library ideas.  The only
first true example `f^Xy` could be done as `a?XntY` with a two
character delta...which is not that bad for how rarely it comes
up.  It also frees up a bad use of a caret, which needs to be
given back to escaping in the language (likely).

(In general, mu-library entities are going to have a much
stronger burden of proof on their existence, henceforth.)