UN => UL for UNLESS, UN for UNION, sets
UO is UNLESS/ONLY at the moment, which may or may not
be important...but UnLess is a bit better anyway, and UN for
UNION seems more fitting.

The set operators are slated to become more expansive and
work as the prefix forms of the bitwise operators for things
like integers and such.