Switch shorthands to use lowercase
For the moment, Ren-C is case-sensitive when it comes to word lookup.

Originally this started while exploring ways of efficiently encoding
"virtual binding", but the original motivation went away.  However,
the question of needing to be able to be case-sensitive came up due
to the small number of identifiers...making it tempting to allow things
like M to represent a module and `m` to represent a different variable.

No hard decisions have been made on this.  But the reason for the
capitalization in the Rebmu definition file was purely visual.

This goes to all lowercase.