Modularize the files
The fledgling module system implements the new concept of isolation.
While the idea hasn't been fully ironed out, the basics now implement
the idea of a "Sea of Words" that separates the idea of word usage
from there needing to be a variable definition for that word.

This makes it possible to actually load and run the scripts under the
new module rules.