Batch of Miscellaneous Changes
The previous 14 commits were attempts to break down what had been
changed in a local copy of Rebmu to get it running under Ren-C, in
order to understand what had actually happened.  (The changes were
made years ago and not broken intelligently into commits.)

This goes ahead and commits the remainder of the changes now that
stock has been taken of what was different by virtue of the partial
commits.