Update uses of SWITCH/DEFAULT and TYPE-OF/WORD
These no longer exist in the default library.

SWITCH now works with ELSE (and can FAIL an evaluative clause in its
body).

TYPE OF is written with a hyphen and there is no /WORD refinement,
which is usually not a problem given SWITCH's evaluativeness.

FAIL replaces THROW for error raising (it used to be DO MAKE ERROR! was
the right way, but FAIL is more succinct).