Add REBMU/OUTPUT to implicitly print output
Some code golf challenges are phrased as "write a function that..."
is intended to calculate a product and printing the output isn't
necessary (e.g. it is enough if it is displayed from a REPL implicitly
as an evaluative result).

This adds a refinement that helps test a program written in that style
by implicilty printing the evaluative product to stdout.